Rename lib/build.rs to lib/builder.rs.
Since build.rs has a fairly special meaning to Cargo if it's at the repo's top
level, having a sub-module called build.rs is a recipe for confusion. Renaming
it to builder.rs should make the difference more obvious.
